Move things from `traits` module to `types` as well
